Fixing issue with Chrome on Android on certain Samsung devices (such as S5) where when your entered text hits a placeholder, the cursor/caret position is not jumped forward (over) the placeholder.
For example, with a phone mask of (___)___-____ . When you get to the 4th position, it should auto jump to the 5th position to skip over the 2nd parentheses. It wasn't in Chrome on Android on Samsung.

This was because the keypress event doesn't exist in that version of Chrome, and the androidInputEvent() was used instead in that configuration.

When setting the caret position, it seems we need to use the $.proxy pattern to give the browser time to correctly place the caret. The keypress event was doing this, but the androidInputEvent() was not. Wrapping the call to input.caret() with the $.proxy call resolved the caret positioning issue.